Move edit prediction metrics into shared crate
