+++ crates/edit_prediction_metrics/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "edit_prediction_metrics"
+version = "0.1.0"
+publish.workspace = true
+edition.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/edit_prediction_metrics.rs"
+
+[dependencies]
+language.workspace = true
+serde.workspace = true
+similar = "2.7.0"
+tree-sitter.workspace = true
+zeta_prompt.workspace = true
+
+[dev-dependencies]
+indoc.workspace = true
+pretty_assertions.workspace = true
